dto class 자료형 변경

+++ src/main/java/com/yet/project/domain/user/SocialLogin.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class SocialLogin {
-    Integer socialId;
+    Long socialId;
     String name;
     String nameKor;
 }
